Create CH1_PracticeProgram4.cpp
Problem Solving With C++
Chapter 1 / Introduction to Computer and C++ Programming
Practice Program 3

Modify the C++ program that you entered in Practice Program 1.
Change the multiplication sign * in your C++ program to an addition sign +.
Recompile and run the changed program.
Notice that the program compiles and runs perfectly fine, but the output is incorrect. That is because this modification is a logic error.