Update CH2_ProgrammingProject2.cpp
Problem Solving With C++
Chapter 2 / C++ Basics
Programming Project 2

Workers at a particular company have won a 7.6% pay increase retroactive for 6 months. Write a program that takes an employee’s previous annual salary as input, and outputs the amount of retroactive pay due the employee, the new annual salary, and the new monthly salary. Your program should allow the calculation to be repeated as often as the user wishes.