Create CH2_ProgrammingProject7.cpp
Problem Solving With C++
Chapter 2 / C++ Basics
Programming Project 7

It is difficult to make a budget that spans several years because prices are not stable. If your company needs 200 pencils per year, you cannot simply use this year’s price as the cost of pencils 2 years from now. Because of inflation, the cost is likely to be higher than it is today. Write a program to gauge the expected cost of an item in a specified number of years. The program asks for the cost of the item, the number of years from now that the item will be purchased, and the rate of inflation. The program then outputs the estimated cost of the item after the specified period. Have the user enter the inflation rate as a percentage, like 5.6 (per cent).